+++ src/management-system-v2/lib/custom-links/custom-link.ts
@@ -1,0 +1,10 @@
+// TODO: check type with zod
+export type CustomNavigationLink = {
+  name: string;
+  icon: string;
+  address: string;
+  topic?: string;
+  showStatus: boolean;
+  clickable: boolean;
+  position: 'top' | 'bottom';
+};
